primeiras alterações  para implementação do frontend

+++ rust/src/models/cliente.rs
@@ -15,7 +15,7 @@ pub struct Endereco {
 
 #[derive(Default)]
 pub struct Cliente {
-    pub id: usize,
+    pub id: i64,
     pub nome: String,
     pub cadastro: String,
     pub endereco: Endereco,
