+++ rust/src/models/pedido.rs
@@ -1,23 +1,23 @@
 #[derive(Default,Clone)]
 pub struct Produto {
-    pub id_merc: usize,
+    pub id_merc: i64,
     pub nome_merc: String,
-    pub quantidade_est: usize
+    pub quantidade_est: i64
 }
 
 #[derive(Default, Clone)]
 pub struct ItemPedido {
-    pub id_item: usize,  
+    pub id_item: i64,  
     pub produto_nome: String,
-    pub quantidade: usize,
+    pub quantidade: i64,
     pub valor: f64,
     pub subtotal: f64,  
 }
 
 #[derive(Default, Clone)]
 pub struct Pedido {
-    pub id: usize,
-    pub cliente_id: usize, 
+    pub id: i64,
+    pub cliente_id: i64, 
     pub itens: Vec<ItemPedido>,
     pub valor_total: f64,
     pub data: String,
